requirements.txt
# Core Data Processing
pandas>=2.3.1
numpy<2.0.0
PyYAML>=6.0

# Korean Stock Market Data
pykrx>=1.0.51
finance-datareader>=0.9.0

# Web Scraping & HTTP
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
mplfinance>=0.12.10b0

# Machine Learning & NLP
scikit-learn>=1.3.0
torch>=1.13.0,<2.0.0
sentence-transformers>=2.2.0
bertopic>=0.15.0
konlpy>=0.6.0
umap-learn>=0.5.0
hdbscan>=0.8.0

# Time Series Analysis & Technical Indicators
backtesting>=0.3.0

# Scheduling & Automation
APScheduler>=3.10.0

# Telegram Bot
python-telegram-bot>=20.0

# Web Driver
webdriver-manager>=4.0.0

# Progress Bars
tqdm>=4.65.0

# Word Processing
gensim>=4.3.3

# Additional ML Libraries
transformers>=4.30.0

# Data Manipulation
openpyxl>=3.1.0

# Date/Time Processing
python-dateutil>=2.8.0

# Logging Enhancement
colorlog>=6.7.0

# HTTP Client
httpx>=0.24.0

# Configuration Management
python-dotenv>=1.0.0

# JSON Processing
ujson>=5.8.0

# Performance Monitoring
psutil>=5.9.0

# Korean Text Processing
mecab-python3>=1.0.0

# Additional Utilities
glob2>=0.7.0
zipfile36>=0.1.0

# FastAPI & Web Server
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6 